pyproject.toml
[build-system]
requires      = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "geoserver-pyadm"
version = "1.0.0"
description = "Geoserver Python client"
readme = "README.md"
authors = [{ name = "Michael Chin", email = "michael.chin@sydney.edu.au" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["geoserver", "RESTful", "API"]
dependencies = [
    "python-dotenv",
    "requests",
]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = ["black", "bumpver", "isort", "pip-tools", "pytest"]

[project.urls]
Homepage = "https://github.com/michaelchin/geoserver-pyadm"

[project.scripts]
geoserver-pyadmin = "geoserver_pyadm.__main__:main"